Support for syncing timers across tabs
If a user had two tabs open, and left one in the background while
working in the other, the background tab could time out and log the user
out, which would cause a logout in the active tab as well.

This commit adds an optional syncronization mechanism for a timer, which
will use localStorage to send and receive activity events. When a tab is
active, it'll set a key in localStorage, causing all other tabs to
receive an event and reset theier activity timeouts.

Bumped version to 1.1.0